# Used a cheaper type check in toMap function, to improve performance