--- a/src/objecttomap.js
+++ b/src/objecttomap.js
@@ -13,6 +13,8 @@
  *		const map = objectToMap( { 'foo': 1, 'bar': 2 } );
  *		map.get( 'foo' ); // 1
  *
+ * **Note**: For mixed data (`Object` or `Iterable`) there's a dedicated {@link module:utils/tomap~toMap} function.
+ *
  * @param {Object} obj Object to transform.
  * @returns {Map} Map created from object.
  */