@oleq 2020-02-05
Guys, make sure these PRs are backed by noticeable performance gain confirmed in Chrome Dev Tools. For instance, do some [heavy editor startup](https://github.com/ckeditor/ckeditor5-utils/pull/322#issuecomment-581962587) check before and after.